refactor(cli): A.3.4 业务 inline observe 4 sub 到 oclif Command
observe.ts 4 个 sub（default health / ingest / show / inbox）业务整段 inline 到
src/cli/oclif/commands/observe/*.ts。删 src/cli/commands/observe.ts。

runObserveInbox 作 module-level helper export from oclif observe/inbox.ts，
test/cli/observe.test.ts in-process import 验证 by-skill 聚合。其它 3 个 sub
（health / ingest / show）业务 inline 到 run() body，不 export。

3 个 helper（pickSkillCount / pickSkillString / pickSkillToolCounts）跟随
runObserveInbox 移到 observe/inbox.ts file-local。

yarn lint && build && test 1494 全绿。
