+++ src/cli/oclif/commands/observe/inbox.ts
@@ -1,8 +1,97 @@
-// oclif 版 observe inbox — 透传 argv 给生产 executeInbox()。
-
+import { resolve } from 'node:path';
 import { Command, Flags } from '@oclif/core';
 import { bilingual } from '../../i18n.js';
 import { runLegacyCommand } from '../../run-legacy.js';
+import { type CliLang } from '../../../i18n.js';
+import type { ObserveInboxArgs, ObserveInboxFlags } from '../../../types/cmd-flags.js';
+
+function pickSkillCount(value: Record<string, number> | undefined, skillName: string): Record<string, number> | undefined {
+  if (!value || value[skillName] == null) return undefined;
+  return { [skillName]: value[skillName] };
+}
+
+function pickSkillString(value: Record<string, string> | undefined, skillName: string): Record<string, string> | undefined {
+  if (!value || value[skillName] == null) return undefined;
+  return { [skillName]: value[skillName] };
+}
+
+function pickSkillToolCounts(value: Record<string, Record<string, number>> | undefined, skillName: string): Record<string, Record<string, number>> | undefined {
+  if (!value || value[skillName] == null) return undefined;
+  return { [skillName]: value[skillName] };
+}
+
+// runObserveInbox export:test/cli/observe.test.ts in-process import 验证 by-skill 聚合行为。
+export async function runObserveInbox(
+  _args: ObserveInboxArgs,
+  flags: ObserveInboxFlags,
+  lang: CliLang,
+): Promise<void> {
+  const { queryObservationInbox, selectExploreInboxItems, loadLatestObservationInboxReports, summarizeObservationInboxBySkill, DEFAULT_OBSERVATIONS_DIR } = await import('../../../../observability/inbox.js');
+  const dir = resolve(flags['input-dir'] || DEFAULT_OBSERVATIONS_DIR);
+  let items = queryObservationInbox(dir);
+  if (flags.skill) {
+    items = items.filter((item) => item.skillName === flags.skill);
+  }
+  if (flags['by-skill']) {
+    const reports = flags.skill
+      ? loadLatestObservationInboxReports(dir).map((report) => ({
+        ...report,
+        meta: {
+          ...report.meta,
+          skillInvocationCounts: pickSkillCount(report.meta.skillInvocationCounts, String(flags.skill)),
+          skillSessionCounts: pickSkillCount(report.meta.skillSessionCounts, String(flags.skill)),
+          skillInvocationLastSeen: pickSkillString(report.meta.skillInvocationLastSeen, String(flags.skill)),
+          skillToolCallCounts: pickSkillToolCounts(report.meta.skillToolCallCounts, String(flags.skill)),
+        },
+      }))
+      : loadLatestObservationInboxReports(dir);
+    const rows = summarizeObservationInboxBySkill(items, reports);
+    if (flags.json) {
+      console.log(JSON.stringify({ kind: 'observe-inbox-by-skill', rows }, null, 2));
+      return;
+    }
+    if (rows.length === 0) {
+      console.log(lang === 'zh' ? 'observe inbox 为空' : 'observe inbox is empty');
+      return;
+    }
+    console.log(lang === 'zh' ? 'observe inbox by skill:' : 'observe inbox by skill:');
+    for (const row of rows) {
+      console.log(`- ${row.skillName} invocations=${row.invocationCount} sessions=${row.sessionCount} processFindings=${row.observationCount} high=${row.highCount} medium=${row.mediumCount} low=${row.lowCount} noise=${row.noiseCount}${row.latestSeen ? ` latest=${row.latestSeen}` : ''}`);
+    }
+    return;
+  }
+  if (flags.explore) {
+    const n = Math.max(1, Number(flags.explore) || 10);
+    items = selectExploreInboxItems(items, n, flags['include-noise']);
+  } else {
+    const limit = Math.max(1, Number(flags.limit ?? 20) || 20);
+    items = items.slice(0, limit);
+  }
+  if (flags.json) {
+    console.log(JSON.stringify({ kind: 'observe-inbox-query', items }, null, 2));
+    return;
+  }
+  if (items.length === 0) {
+    console.log(lang === 'zh' ? 'observe inbox 为空' : 'observe inbox is empty');
+    return;
+  }
+  console.log(lang === 'zh' ? 'observe inbox:' : 'observe inbox:');
+  for (const item of items) {
+    const evidence = item.evidence.query || item.evidence.path || item.evidence.assistantSnippet || item.evidence.outputSnippet || '';
+    const artifactVersion = item.artifactVersion === 'unknown' ? '⚠ unknown' : item.artifactVersion;
+    console.log(`- [${item.severity}] (${item.sourceKind}) ${item.skillName} ${item.signalType}/${item.signalSubtype} x${item.occurrences} confidence=${item.confidence.toFixed(2)} attribution=${item.attributionConfidence.toFixed(2)}`);
+    console.log(`  lastSeen=${item.lastSeen} version=${artifactVersion}`);
+    console.log(`  reason=${item.severityReasonCode ?? 'unknown'}`);
+    if (evidence) console.log(`  evidence=${evidence.slice(0, 180)}`);
+  }
+  console.log('');
+  console.log(lang === 'zh'
+    ? 'Tip: omk observe inbox --explore 10  # 抽样查看 medium/low 长尾'
+    : 'Tip: omk observe inbox --explore 10  # sample medium/low long-tail items');
+  console.log(lang === 'zh'
+    ? 'Tip: omk observe inbox --explore 10 --include-noise  # 显式包含 noise 桶'
+    : 'Tip: omk observe inbox --explore 10 --include-noise  # explicitly include the noise bucket');
+}
 
 export default class ObserveInbox extends Command {
   static description = bilingual({
@@ -57,7 +146,6 @@ export default class ObserveInbox extends Command {
     const { args, flags } = await this.parse(ObserveInbox);
     const lang = (flags.lang ?? 'zh') as 'zh' | 'en';
     await runLegacyCommand(this, async () => {
-      const { runObserveInbox } = await import('../../../commands/observe.js');
       await runObserveInbox(args as Record<string, never>, { ...flags, lang }, lang);
     });
   }
